Extract display helpers
